ingestion/iot_streaming/requirements.txt
# Dépendances pour le connecteur MQTT → Kinesis Bracelet IoT Kidjamo

# AWS SDK et IoT
boto3>=1.34.0
awsiotsdk>=1.11.0
botocore>=1.34.0

# MQTT et communication
paho-mqtt>=1.6.0
awscrt>=0.16.0

# Traitement des données
pandas>=2.0.0
numpy>=1.24.0

# Sérialisation et validation
pydantic>=2.0.0
jsonschema>=4.0.0

# Logging et monitoring
structlog>=23.0.0
python-json-logger>=2.0.0

# Async et concurrence
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Utilitaires
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0

# Tests et développement (optionnel)
pytest>=7.0.0
pytest-asyncio>=0.21.0
fakeredis>=2.0.0

# Sécurité
cryptography>=41.0.0
certifi>=2023.0.0
